backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
pymysql==1.1.1
mysql-connector-python==9.4.0
cryptography>=42.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
pandas==2.3.1
openpyxl==3.1.5
reportlab==4.2.5
matplotlib==3.9.2
